requirements.txt
# 新闻聚合器依赖包
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedgen>=1.0.0
Jinja2>=3.1.2
PyYAML>=6.0
aiohttp>=3.8.0
aiofiles>=23.1.0
python-dateutil>=2.8.0
lxml>=4.9.0
html5lib>=1.1
urllib3>=2.0.0